* src/tap/cable/*.c: convert return values, error handling, printfs   * src/**: globally remove \n or !\n from error detail msgs; sanitize     error prints a tiny bit
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@1594 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ urjtag/src/global/parse.c
@@ -65,7 +65,8 @@ urj_parse_line (urj_chain_t *chain, char *line)
     sline = malloc (l + 1);
     if (sline == NULL)
     {
-        urj_error_set (URJ_ERROR_OUT_OF_MEMORY, "malloc(%d) fails", l + 1);
+        urj_error_set (URJ_ERROR_OUT_OF_MEMORY, "malloc(%zd) fails",
+                       (size_t) (l + 1));
         return URJ_STATUS_FAIL;
     }
 
@@ -104,7 +105,7 @@ urj_parse_line (urj_chain_t *chain, char *line)
     if (a == NULL)
     {
         urj_error_set (URJ_ERROR_OUT_OF_MEMORY, "malloc(%zd) fails",
-                       (tcnt + 1) * sizeof (char *));
+                       (size_t) ((tcnt + 1) * sizeof (char *)));
         return URJ_STATUS_FAIL;
     }
